For wasm

+++ rust/utils/src/dataset/mnist.rs
@@ -8,14 +8,14 @@ use std::path::Path;
 
 pub fn load_mnist<P: AsRef<Path>>(
     path: Vec<P>,
-) -> ((Array2<f64>, Array2<f64>), (Array2<f64>, Array2<f64>)) {
+) -> ((Array2<f32>, Array2<f32>), (Array2<f32>, Array2<f32>)) {
     let mut path_iter = path.into_iter();
-    let (train_x, num_image_train): (Vec<f64>, usize) = load_images(path_iter.next().unwrap());
-    let (train_y, num_label_train): (Vec<f64>, usize) = load_labels(path_iter.next().unwrap());
+    let (train_x, num_image_train): (Vec<f32>, usize) = load_images(path_iter.next().unwrap());
+    let (train_y, num_label_train): (Vec<f32>, usize) = load_labels(path_iter.next().unwrap());
 
-    let (test_x, num_image_test): (Vec<f64>, usize) = load_images(path_iter.next().unwrap());
+    let (test_x, num_image_test): (Vec<f32>, usize) = load_images(path_iter.next().unwrap());
 
-    let (test_y, num_label_test): (Vec<f64>, usize) = load_labels(path_iter.next().unwrap());
+    let (test_y, num_label_test): (Vec<f32>, usize) = load_labels(path_iter.next().unwrap());
 
     let as_arr = Array::from_shape_vec; // create am array withe the given shape from a vector
     let x_train = as_arr(Ix2(num_image_train, 28 * 28), train_x).unwrap();
@@ -28,7 +28,7 @@ pub fn load_mnist<P: AsRef<Path>>(
     )
 }
 
-fn load_images<P: AsRef<Path>>(path: P) -> (Vec<f64>, usize) {
+fn load_images<P: AsRef<Path>>(path: P) -> (Vec<f32>, usize) {
     let file = File::open(path).expect("please sure the data file exists");
     let ref mut buf_reader = io::BufReader::new(file);
     let magic = read_be_u32(buf_reader);
@@ -45,11 +45,11 @@ fn load_images<P: AsRef<Path>>(path: P) -> (Vec<f64>, usize) {
 
     let mut buf: Vec<u8> = vec![0 as u8; num_image * rows * cols];
     let _ = buf_reader.read_exact(buf.as_mut());
-    let ret: Vec<f64> = buf.into_iter().map(|x| (x as f64) / 255.).collect();
+    let ret: Vec<f32> = buf.into_iter().map(|x| (x as f32) / 255.).collect();
     (ret, num_image)
 }
 
-fn load_labels<P: AsRef<Path>>(path: P) -> (Vec<f64>, usize) {
+fn load_labels<P: AsRef<Path>>(path: P) -> (Vec<f32>, usize) {
     let ref mut buf_reader =
         io::BufReader::new(File::open(path).expect("please sure label data exists"));
     let magic = read_be_u32(buf_reader);
@@ -61,7 +61,7 @@ fn load_labels<P: AsRef<Path>>(path: P) -> (Vec<f64>, usize) {
     let num_label = read_be_u32(buf_reader) as usize;
     let mut buf: Vec<u8> = vec![0 as u8; num_label];
     let _ = buf_reader.read_exact(buf.as_mut());
-    let ret: Vec<f64> = buf.into_iter().map(|x| x as f64).collect();
+    let ret: Vec<f32> = buf.into_iter().map(|x| x as f32).collect();
     (ret, num_label)
 }
